fixed compile warning

+++ source/opengl/gles.cpp
@@ -449,8 +449,8 @@ void gl_common_EglCreateContext(CPU* cpu) {
     EAX = KNativeSystem::getOpenGL()->glCreateContext(thread, cfg->glPixelFormat, major, minor, profile, 0, ARG3);
     if (eglLog()) {
 #ifdef __EMSCRIPTEN__
-        klog_fmt("boxedwine EGL: eglCreateContext config=0x%x resolved=0x%x version=%d.%d profile=0x%x share=%u -> %u webgl=%d",
-                ARG2, config, major, minor, profile, ARG3, EAX, emscripten_webgl_get_current_context());
+        klog_fmt("boxedwine EGL: eglCreateContext config=0x%x resolved=0x%x version=%d.%d profile=0x%x share=%u -> %u webgl=%lu",
+                ARG2, config, major, minor, profile, ARG3, EAX, static_cast<unsigned long>(emscripten_webgl_get_current_context()));
 #else
         klog_fmt("boxedwine EGL: eglCreateContext config=0x%x resolved=0x%x fb=%u version=%d.%d profile=0x%x share=%u -> %u", ARG2, config, cfg->fbId, major, minor, profile, ARG3, EAX);
 #endif
@@ -557,8 +557,8 @@ void gl_common_EglMakeCurrent(CPU* cpu) {
     }
     if (eglLog()) {
 #ifdef __EMSCRIPTEN__
-        klog_fmt("boxedwine EGL: eglMakeCurrent draw=%u read=%u ctx=%u drawable=%u -> %u webgl=%d",
-                draw, read, ctx, d ? d->id : 0, EAX, emscripten_webgl_get_current_context());
+        klog_fmt("boxedwine EGL: eglMakeCurrent draw=%u read=%u ctx=%u drawable=%u -> %u webgl=%lu",
+                draw, read, ctx, d ? d->id : 0, EAX, static_cast<unsigned long>(emscripten_webgl_get_current_context()));
 #else
         klog_fmt("boxedwine EGL: eglMakeCurrent draw=%u read=%u ctx=%u drawable=%u -> %u", draw, read, ctx, d ? d->id : 0, EAX);
 #endif
